+++ Backend/Reservas/src/main/java/com/reserva/Reserva/repositories/ReservaRepository.java
@@ -1,8 +1,9 @@
 package com.reserva.Reserva.repositories;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.reserva.Reserva.Model.entities.Reserva;
 
-public interface ReservaRepository extends JpaRepository<Reserva, Long> {
-    
+public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
+
 }
